사냥터 구현 - 밸런싱 수치 적용

+++ Assets/Scripts/Data/Character.cs
@@ -8,7 +8,7 @@ public class CharInfo
     public int Char_Level = 1;
     public int Char_Str = 1;
     public int Char_Con = 1;
-    public int Char_HP = 100;
+    public float Char_HP = 100f;
     public int Char_Exp = 100;
 }
