+++ Assets/Scripts/Data/PlayerStat.cs
@@ -11,7 +11,7 @@ public class PlayerStat
     public int PC_Con;
     public int PC_Exp;
     public int PC_UpExp;
-    public int PC_MaxHP;
+    public float PC_MaxHP;
     public string PC_Gold;
     public int PC_WpnID;
     public int PC_WpnEct;
